Renamed MetricRunner to ScenarioRunner
