+++ frontend/cluster-status.html
@@ -33,7 +33,7 @@
     <div class="main ui container" id="main">
         <h1 class="ui header float left">
             <a id="menu-toggle" class="opened"><i class="bars bordered inverted left icon opened"></i></a>
-            MetricRunner - Cluster Status
+            ScenarioRunner - Cluster Status
         </h1>
         <div class="ui full-width-card segment card">
             <div class="content">
