+++ frontend/index.html
@@ -49,18 +49,18 @@
             </div>
         </div>
         <div class="ui cards link">
-            <div class="ui card metric-runner index-connected-tools">
+            <div class="ui card scenario-runner index-connected-tools">
                 <div class="image" style="background: #3d936f; padding: 30px 0;">
                     <a href="/runs.html" class="project-item"><i class="tachometer alternate icon project-item-white" style="margin: 0px auto;"></i></a>
                 </div>
                 <div class="content">
-                    <a class="header">MetricRunner</a>
+                    <a class="header">ScenarioRunner</a>
                     <div class="meta">
                         <span class="date">Measure software for energy and CO₂</span>
                     </div>
                     <div class="description">
-                        <p>Current Runs in DB: <b><span id="metric-runner-count">N/A</span></b></p>
-                        <p>Oldest Run in DB: <b><span id="metric-runner-oldest">N/A</span></b></p>
+                        <p>Current Runs in DB: <b><span id="scenario-runner-count">N/A</span></b></p>
+                        <p>Oldest Run in DB: <b><span id="scenario-runner-oldest">N/A</span></b></p>
                     </div>
                     <br> &raquo; Additional Info ... etc. ???
                 </div>
