+++ frontend/js/helpers/config.js.example
@@ -2,7 +2,7 @@
 API_URL = "__API_URL__"
 METRICS_URL = "__METRICS_URL__"
 
-ACTIVATE_METRIC_RUNNER = __ACTIVATE_METRIC_RUNNER__;
+ACTIVATE_SCENARIO_RUNNER = __ACTIVATE_SCENARIO_RUNNER__;
 ACTIVATE_CARBON_DB = __ACTIVATE_CARBON_DB__;
 ACTIVATE_ECO_CI = __ACTIVATE_ECO_CI__;
 ACTIVATE_POWER_HOG = __ACTIVATE_POWER_HOG__;
