+++ frontend/js/index.js
@@ -1,11 +1,11 @@
 (async () => {
-    if (ACTIVATE_METRIC_RUNNER === true) {
-        document.querySelectorAll('.metric-runner').forEach(el => el.style.setProperty("display", "block", "important"))
+    if (ACTIVATE_SCENARIO_RUNNER === true) {
+        document.querySelectorAll('.scenario-runner').forEach(el => el.style.setProperty("display", "block", "important"))
         try {
             const api_data = await makeAPICall(`/v1/insights`)
-            document.querySelector('#metric-runner-count').innerText = api_data.data[0];
-            document.querySelector('#metric-runner-oldest').innerText = api_data.data[1];
-        } catch (err) { showNotification(`Could not get MetricRunner stats from API`, err) }
+            document.querySelector('#scenario-runner-count').innerText = api_data.data[0];
+            document.querySelector('#scenario-runner-oldest').innerText = api_data.data[1];
+        } catch (err) { showNotification(`Could not get ScenarioRunner stats from API`, err) }
     }
 
     if (ACTIVATE_ECO_CI === true) {
