+++ frontend/js/metric-runner.js
@@ -47,25 +47,25 @@ async function getRepositories(sort_by = 'date') {
 (async () => {
     document.querySelector('#home-toggle-button').addEventListener('click', el => {
         if (el.currentTarget.innerText === 'Switch to repository view') {
-            document.querySelector('h1.ui.header span').innerText = 'MetricRunner - Repositories';
-            localStorage.setItem('metric_runner_data_shown', 'repositories');
+            document.querySelector('h1.ui.header span').innerText = 'ScenarioRunner - Repositories';
+            localStorage.setItem('scenario_runner_data_shown', 'repositories');
             window.location.reload();
         } else {
-            document.querySelector('h1.ui.header span').innerText = 'MetricRunner - Last 50 Runs';
-            localStorage.setItem('metric_runner_data_shown', 'last_runs');
+            document.querySelector('h1.ui.header span').innerText = 'ScenarioRunner - Last 50 Runs';
+            localStorage.setItem('scenario_runner_data_shown', 'last_runs');
             window.location.reload();
         }
     });
 
-    if (localStorage.getItem('metric_runner_data_shown') === 'repositories') {
+    if (localStorage.getItem('scenario_runner_data_shown') === 'repositories') {
         document.querySelector('#home-toggle-button').innerText = 'Switch to last runs view';
-        document.querySelector('h1.ui.header span').innerText = 'MetricRunner - Repositories';
-        document.querySelectorAll('.metric-runner-repositories').forEach(el => el.style.visibility = 'visible');
-        document.querySelector('#metric-runner-runs-description')?.remove();
+        document.querySelector('h1.ui.header span').innerText = 'ScenarioRunner - Repositories';
+        document.querySelectorAll('.scenario-runner-repositories').forEach(el => el.style.visibility = 'visible');
+        document.querySelector('#scenario-runner-runs-description')?.remove();
         sortDate();
     } else {
-        document.querySelectorAll('.metric-runner-runs').forEach(el => el.style.visibility = 'visible');
-        document.querySelector('#metric-runner-repositories-description')?.remove();
+        document.querySelectorAll('.scenario-runner-runs').forEach(el => el.style.visibility = 'visible');
+        document.querySelector('#scenario-runner-repositories-description')?.remove();
         getRunsTable($('#runs-table'), `/v1/runs?${getFilterQueryStringFromURI()}&limit=50`)
     }
 })();
