+++ frontend/projects.html
@@ -10,7 +10,7 @@
     <link rel="icon" href="/images/favicon-16x16.png" sizes="16x16">
     <link rel="icon" href="/images/favicon-32x32.png" sizes="32x32">
     <!-- Site Properties -->
-    <title>Green Metrics Tool - MetricRunner Projects</title>
+    <title>Green Metrics Tool - ScenarioRunner Projects</title>
     <meta name="referrer" content="no-referrer-when-downgrade" />
     <meta name="description" content="Measuring the energy and carbon footprint of your software" />
     <script src="/dist/js/jquery.min.js" defer></script>
@@ -19,7 +19,7 @@
     <script src="/dist/js/toast.min.js" defer></script>
     <script src="/js/helpers/config.js" defer></script>
     <script src="/js/helpers/main.js" defer></script>
-    <script src="/js/metric-runner-projects.js" defer></script>
+    <script src="/js/scenario-runner-projects.js" defer></script>
 
     <link rel="stylesheet" type="text/css" class="ui" href="/dist/css/semantic_reduced.min.css">
     <link rel="stylesheet" type="text/css" href="/css/green-coding.css">
@@ -29,20 +29,20 @@
     <div class="main ui container" id="main">
         <h1 class="ui header float left">
             <a id="menu-toggle" class="opened"><i class="bars bordered inverted left icon openend"></i></a>
-            MetricRunner - Projects
+            ScenarioRunner - Projects
         </h1>
         <div class="ui full-width-card segment card">
             <div class="content">
-                <div class="header">MetricRunner Projects Overview</div>
+                <div class="header">ScenarioRunner Projects Overview</div>
                 <div class="description">
                     <i class="close icon"></i>
-                    <p>This page shows you all projects that are monitored via MetricRunner at the moment</p>
+                    <p>This page shows you all projects that are monitored via ScenarioRunner at the moment</p>
                     <p>The score cards show you the trends on specific key metrics and also the monitoring interval / last date measured.</p>
                     <p>When you click in the badges you will get to the timeline view, where you can see all badges and more charts and metrics.</p>
                 </div>
             </div>
         </div>
-        <div id="metric-runner-projects"class="ui cards"></div>
+        <div id="scenario-runner-projects"class="ui cards"></div>
     </div>
 </body>
 </html>
