+++ frontend/request.html
@@ -10,7 +10,7 @@
     <link rel="icon" href="/images/favicon-16x16.png" sizes="16x16">
     <link rel="icon" href="/images/favicon-32x32.png" sizes="32x32">
     <!-- Site Properties -->
-    <title>Green Metrics Tool - MetricRunner Submit Software</title>
+    <title>Green Metrics Tool - ScenarioRunner Submit Software</title>
     <meta name="referrer" content="no-referrer-when-downgrade" />
     <meta name="description" content="Submit a software for energy measurement to the Green Metrics Tool" />
     <script src="/dist/js/jquery.min.js" defer></script>
@@ -27,17 +27,17 @@
     <div class="main ui container" id="main">
         <h1 class="ui header float left">
             <a id="menu-toggle" class="opened"><i class="bars bordered inverted left icon opened"></i></a>
-            MetricRunner - Submit Software
+            ScenarioRunner - Submit Software
         </h1>
         <div class="ui full-width-card segment card">
             <div class="content">
                 <div class="header">Measuring software introduction</div>
                 <div class="description">
                     <i class="close icon"></i>
-                    <p>The measurement cluster behind MetricRunner can measure your software either in a one-off run or in a continuous monitoring.</p>
+                    <p>The measurement cluster behind ScenarioRunner can measure your software either in a one-off run or in a continuous monitoring.</p>
                     <p>In the box below please put your Github/Gitlab/Bitbucket-Repo URL and an E-mail where we send the link to the statistics.</p>
                     <p>Your run will be put in a queue and processed in the next free slot.</p>
-                    <p>MetricRunner expects a schema-correct
+                    <p>ScenarioRunner expects a schema-correct
                         <a href="https://docs.green-coding.io/docs/measuring/usage-scenario/">usage_scenario.yml</a> file in your repository root. Please check the <a href="https://github.com/green-coding-solutions/simple-example-application">Demo
                             Software Repository</a> for example and documentation.<br/>
                         Override the usage scenario file name or provide relative path if it is not in the repository root with the optional filename field.
