+++ frontend/timeline.html
@@ -10,7 +10,7 @@
     <link rel="icon" href="/images/favicon-16x16.png" sizes="16x16">
     <link rel="icon" href="/images/favicon-32x32.png" sizes="32x32">
     <!-- Site Properties -->
-    <title>Green Metrics Tool - MetricRunner Timeline</title>
+    <title>Green Metrics Tool - ScenarioRunner Timeline</title>
     <meta name="referrer" content="no-referrer-when-downgrade" />
     <meta name="description" content="Measuring the energy and carbon footprint of your software" />
     <script src="/dist/js/jquery.min.js" defer></script>
@@ -31,14 +31,14 @@
     <script src="/js/helpers/metric-boxes.js" defer></script>
     <script src="/js/helpers/converters.js" defer></script>
     <script src="/js/helpers/charts.js" defer></script>
-    <script src="/js/metric-runner-timeline.js" defer></script>
+    <script src="/js/scenario-runner-timeline.js" defer></script>
     <link rel="stylesheet" type="text/css" class="ui" href="/dist/css/semantic_reduced.min.css">
     <link rel="stylesheet" type="text/css" href="/css/green-coding.css">
 </head>
 <body class="preload">
     <gmt-menu></gmt-menu>
     <div class="main ui container" id="main">
-        <h1 class="ui header float left"><a id="menu-toggle" class="opened"><i class="bars bordered inverted left icon opened"></i></a> Green Metrics Tool - MetricRunner Timeline</h1>
+        <h1 class="ui header float left"><a id="menu-toggle" class="opened"><i class="bars bordered inverted left icon opened"></i></a> Green Metrics Tool - ScenarioRunner Timeline</h1>
         <div class="ui full-width-card card">
             <div class="content">
                 <div class="description run-data-container">
@@ -46,7 +46,7 @@
                         <h3>Timeline View Details</h3>
                     </a></div>
                     <h4>What is Timeline View?</h4>
-                    <p>Timeline View shows all the measurements we have for a specific MetricRunner project over time. <br>This allows you to inspect if a software has an upward or downward trend in energy, memory, CO2 etc.</p>
+                    <p>Timeline View shows all the measurements we have for a specific ScenarioRunner project over time. <br>This allows you to inspect if a software has an upward or downward trend in energy, memory, CO2 etc.</p>
                     <p>If you want to change the graphs you are seeing, please click on <b>Customize</b>. The charts will change on refresh.</p>
                     <div class="ui top secondary menu">
                         <a class="item active" data-tab="one">Badges</a>
